Add files via upload
Runs properly. Fixed bugs created with changes to source and header files.
Displays a menu on the right side of the screen. There is a button the user can press when ready to play the game that changes the game state and hides the menu and button.
Uses sprites/textures for the menu and button.