Add files via upload
The LevelSelect class displays a single level selector sprite, Level 1, for the time being.
The MenuView class loads the LevelSelect class passing the renderwindow and the game state.
The LevelSelect class now handles mouse input. If a mouse click is registered on the level selector sprite then the game state is changed to Loading.
The GreatEscape.cpp file will handle loading methods based on the game state value.