Add files via upload
Created a sprite sheet for the level backgrounds and for the small/stolen objects.
I updated the xml to reference the pixel bounds of each object in the sprite sheet. However, this is not yet incorporated in the game's graphics.

When the player selects a level, the gamestate now changes to STORY.
I created an int variable, levelSelected which holds the level the player clicked in the LevelSelect class.
I created a dialogue class which is called in the main class when the GameState is STORY. 
The Dialogue class currently has level 1 and level 2's story hardcoded and drawn on the screen. The player moves to the level by pressing any key.